complete

+++ taskManager/views.py
@@ -168,53 +168,58 @@ def manage_groups(request):
 
 
 def upload(request, project_id):
+    user = request.user
 
-    if request.method == 'POST':
+    if user.is_authenticated():
+        if request.method == 'POST':
 
-        proj = Project.objects.get(pk=project_id)
-        form = ProjectFileForm(request.POST, request.FILES)
+            proj = Project.objects.get(pk=project_id)
+            form = ProjectFileForm(request.POST, request.FILES)
 
-        if form.is_valid():
-            name = request.POST.get('name', False)
-            upload_path = store_uploaded_file(name, request.FILES['file'])
+            if form.is_valid():
+                name = request.POST.get('name', False)
+                upload_path = store_uploaded_file(name, request.FILES['file'])
 
-            #A1 - Injection (SQLi)
-            curs = connection.cursor()
-            curs.execute(
-                "insert into taskManager_file ('name','path','project_id') values ('%s','%s',%s)" %
-                (name, upload_path, project_id))
+                # A1 - Injection (SQLi)
+                # curs = connection.cursor()
+                # curs.execute(
+                #     "insert into taskManager_file ('name','path','project_id') values ('%s','%s',%s)" %
+                #     (name, upload_path, project_id))
 
-            # file = File(
-            #name = name,
-            #path = upload_path,
-            # project = proj)
+                file = File(
+                    name=name,
+                    path=upload_path,
+                    project=proj
+                )
 
-            # file.save()
+                file.save()
 
-            return redirect('/taskManager/' + project_id +
-                            '/', {'new_file_added': True})
+                return redirect('/taskManager/' + project_id +
+                                '/', {'new_file_added': True})
+            else:
+                form = ProjectFileForm()
         else:
             form = ProjectFileForm()
-    else:
-        form = ProjectFileForm()
-    return render_to_response(
-        'taskManager/upload.html', {'form': form}, RequestContext(request))
+        return render_to_response(
+            'taskManager/upload.html', {'form': form}, RequestContext(request))
 
 # A4: Insecure Direct Object Reference (IDOR)
 
 
 def download(request, file_id):
+    user = request.user
 
-    file = File.objects.get(pk=file_id)
-    abspath = open(
-        os.path.dirname(
-            os.path.realpath(__file__)) +
-        file.path,
-        'rb')
-    response = HttpResponse(content=abspath.read())
-    response['Content-Type'] = mimetypes.guess_type(file.path)[0]
-    response['Content-Disposition'] = 'attachment; filename=%s' % file.name
-    return response
+    if user.is_authenticated():
+        file = File.objects.get(pk=file_id)
+        abspath = open(
+            os.path.dirname(
+                os.path.realpath(__file__)) +
+            file.path,
+            'rb')
+        response = HttpResponse(content=abspath.read())
+        response['Content-Type'] = mimetypes.guess_type(file.path)[0]
+        response['Content-Disposition'] = 'attachment; filename=%s' % file.name
+        return response
 
 
 def download_profile_pic(request, user_id):
@@ -238,85 +243,94 @@ def download_profile_pic(request, user_id):
 
 
 def task_create(request, project_id):
+    user = request.user
 
-    if request.method == 'POST':
+    if user.is_authenticated():
+        if request.method == 'POST':
 
-        proj = Project.objects.get(pk=project_id)
+            proj = Project.objects.get(pk=project_id)
 
-        text = request.POST.get('text', False)
-        task_title = request.POST.get('task_title', False)
-        now = timezone.now()
-        task_duedate = timezone.now() + datetime.timedelta(weeks=1)
-        if request.POST.get('task_duedate') != '':
-            task_duedate = datetime.datetime.fromtimestamp(
-                int(request.POST.get('task_duedate', False)))
+            text = request.POST.get('text', False)
+            task_title = request.POST.get('task_title', False)
+            now = timezone.now()
+            task_duedate = timezone.now() + datetime.timedelta(weeks=1)
+            if request.POST.get('task_duedate') != '':
+                task_duedate = datetime.datetime.fromtimestamp(
+                    int(request.POST.get('task_duedate', False)))
 
-        task = Task(
-            text=text,
-            title=task_title,
-            start_date=now,
-            due_date=task_duedate,
-            project=proj)
+            task = Task(
+                text=text,
+                title=task_title,
+                start_date=now,
+                due_date=task_duedate,
+                project=proj)
 
-        task.save()
-        task.users_assigned = [request.user]
+            task.save()
+            task.users_assigned = [request.user]
 
-        return redirect('/taskManager/' + project_id +
-                        '/', {'new_task_added': True})
-    else:
-        return render_to_response(
-            'taskManager/task_create.html', {'proj_id': project_id}, RequestContext(request))
+            return redirect('/taskManager/' + project_id +
+                            '/', {'new_task_added': True})
+        else:
+            return render_to_response(
+                'taskManager/task_create.html', {'proj_id': project_id}, RequestContext(request))
 
 # A4: Insecure Direct Object Reference (IDOR)
 
 
 def task_edit(request, project_id, task_id):
+    user = request.user
+    if user.is_authenticated():
+        proj = Project.objects.get(pk=project_id)
+        task = Task.objects.get(pk=task_id)
 
-    proj = Project.objects.get(pk=project_id)
-    task = Task.objects.get(pk=task_id)
+        if request.method == 'POST':
 
-    if request.method == 'POST':
+            if task.project == proj:
 
-        if task.project == proj:
+                text = request.POST.get('text', False)
+                task_title = request.POST.get('task_title', False)
+                task_completed = request.POST.get('task_completed', False)
 
-            text = request.POST.get('text', False)
-            task_title = request.POST.get('task_title', False)
-            task_completed = request.POST.get('task_completed', False)
+                task.title = task_title
+                task.text = text
+                task.completed = True if task_completed == "1" else False
+                task.save()
 
-            task.title = task_title
-            task.text = text
-            task.completed = True if task_completed == "1" else False
-            task.save()
-
-        return redirect('/taskManager/' + project_id + '/' + task_id)
-    else:
-        return render_to_response(
-            'taskManager/task_edit.html', {'task': task}, RequestContext(request))
+            return redirect('/taskManager/' + project_id + '/' + task_id)
+        else:
+            return render_to_response(
+                'taskManager/task_edit.html', {'task': task}, RequestContext(request))
 
 # A4: Insecure Direct Object Reference (IDOR)
 
 
 def task_delete(request, project_id, task_id):
-    proj = Project.objects.get(pk=project_id)
-    task = Task.objects.get(pk=task_id)
-    if proj is not None:
-        if task is not None and task.project == proj:
-            task.delete()
+    user = request.user
+    
+    if user.is_authenticated():
+        proj = Project.objects.get(pk=project_id)
+        task = Task.objects.get(pk=task_id)
+        if proj is not None:
+            if task is not None and task.project == proj:
+                task.delete()
 
-    return redirect('/taskManager/' + project_id + '/')
+        return redirect('/taskManager/' + project_id + '/')
 
 # A4: Insecure Direct Object Reference (IDOR)
 
 
 def task_complete(request, project_id, task_id):
-    proj = Project.objects.get(pk=project_id)
-    task = Task.objects.get(pk=task_id)
-    if proj is not None:
-        if task is not None and task.project == proj:
-            task.completed = not task.completed
-            task.save()
+    user = request.user
+    
+    if user.is_authenticated():
+        proj = Project.objects.get(pk=project_id)
+        task = Task.objects.get(pk=task_id)
+        if proj is not None:
+            if task is not None and task.project == proj:
+                task.completed = not task.completed
+                task.save()
 
-    return redirect('/taskManager/' + project_id)
+        return redirect('/taskManager/' + project_id)
 
 
 def project_create(request):
@@ -348,36 +362,41 @@ def project_create(request):
 
 # A4: Insecure Direct Object Reference (IDOR)
 def project_edit(request, project_id):
+    user = request.user
 
-    proj = Project.objects.get(pk=project_id)
+    if user.is_authenticated():
+        proj = Project.objects.get(pk=project_id)
 
-    if request.method == 'POST':
+        if request.method == 'POST':
 
-        title = request.POST.get('title', False)
-        text = request.POST.get('text', False)
-        project_priority = int(request.POST.get('project_priority', False))
-        project_duedate = datetime.datetime.fromtimestamp(
-            int(request.POST.get('project_duedate', False)))
+            title = request.POST.get('title', False)
+            text = request.POST.get('text', False)
+            project_priority = int(request.POST.get('project_priority', False))
+            project_duedate = datetime.datetime.fromtimestamp(
+                int(request.POST.get('project_duedate', False)))
 
-        proj.title = title
-        proj.text = text
-        proj.priority = project_priority
-        proj.due_date = project_duedate
-        proj.save()
+            proj.title = title
+            proj.text = text
+            proj.priority = project_priority
+            proj.due_date = project_duedate
+            proj.save()
 
-        return redirect('/taskManager/' + project_id + '/')
-    else:
-        return render_to_response(
-            'taskManager/project_edit.html', {'proj': proj}, RequestContext(request))
+            return redirect('/taskManager/' + project_id + '/')
+        else:
+            return render_to_response(
+                'taskManager/project_edit.html', {'proj': proj}, RequestContext(request))
 
 # A4: Insecure Direct Object Reference (IDOR)
 
 
 def project_delete(request, project_id):
-    # IDOR
-    project = Project.objects.get(pk=project_id)
-    project.delete()
-    return redirect('/taskManager/dashboard')
+    user = request.user
+
+    if user.is_authenticated():
+        # IDOR
+        project = Project.objects.get(pk=project_id)
+        project.delete()
+        return redirect('/taskManager/dashboard')
 
 # A10: Open Redirect
 
@@ -552,42 +571,48 @@ def note_create(request, project_id, task_id):
 
 
 def note_edit(request, project_id, task_id, note_id):
+    user = request.user
 
-    proj = Project.objects.get(pk=project_id)
-    task = Task.objects.get(pk=task_id)
-    note = Notes.objects.get(pk=note_id)
+    if user.is_authenticated():
 
-    if request.method == 'POST':
+        proj = Project.objects.get(pk=project_id)
+        task = Task.objects.get(pk=task_id)
+        note = Notes.objects.get(pk=note_id)
 
-        if task.project == proj:
+        if request.method == 'POST':
 
-            if note.task == task:
+            if task.project == proj:
 
-                text = request.POST.get('text', False)
-                note_title = request.POST.get('note_title', False)
+                if note.task == task:
 
-                note.title = note_title
-                note.text = text
-                note.save()
+                    text = request.POST.get('text', False)
+                    note_title = request.POST.get('note_title', False)
 
-        return redirect('/taskManager/' + project_id + '/' + task_id)
-    else:
-        return render_to_response(
-            'taskManager/note_edit.html', {'note': note}, RequestContext(request))
+                    note.title = note_title
+                    note.text = text
+                    note.save()
+
+            return redirect('/taskManager/' + project_id + '/' + task_id)
+        else:
+            return render_to_response(
+                'taskManager/note_edit.html', {'note': note}, RequestContext(request))
 
 # A4: Insecure Direct Object Reference (IDOR)
 
 
 def note_delete(request, project_id, task_id, note_id):
-    proj = Project.objects.get(pk=project_id)
-    task = Task.objects.get(pk=task_id)
-    note = Notes.objects.get(pk=note_id)
-    if proj is not None:
-        if task is not None and task.project == proj:
-            if note is not None and note.task == task:
-                note.delete()
+    user = request.user
 
-    return redirect('/taskManager/' + project_id + '/' + task_id)
+    if user.is_authenticated():
+        proj = Project.objects.get(pk=project_id)
+        task = Task.objects.get(pk=task_id)
+        note = Notes.objects.get(pk=note_id)
+        if proj is not None:
+            if task is not None and task.project == proj:
+                if note is not None and note.task == task:
+                    note.delete()
+
+        return redirect('/taskManager/' + project_id + '/' + task_id)
 
 
 def task_details(request, project_id, task_id):
@@ -709,30 +734,33 @@ def profile(request):
 
 @csrf_exempt
 def profile_by_id(request, user_id):
-    user = User.objects.get(pk=user_id)
+    user = request.user
 
-    if request.method == 'POST':
-        form = ProfileForm(request.POST, request.FILES)
-        if form.is_valid():
-            print("made it!")
-            if request.POST.get('username') != user.username:
-                user.username = request.POST.get('username')
-            if request.POST.get('first_name') != user.first_name:
-                user.first_name = request.POST.get('first_name')
-            if request.POST.get('last_name') != user.last_name:
-                user.last_name = request.POST.get('last_name')
-            if request.POST.get('email') != user.email:
-                user.email = request.POST.get('email')
-            if request.POST.get('password'):
-                user.set_password(request.POST.get('password'))
-            if request.FILES:
-                user.userprofile.image = store_uploaded_file(user.username
-                + "." + request.FILES['picture'].name.split(".")[-1], request.FILES['picture'])
-                user.userprofile.save()
-            user.save()
-            messages.info(request, "User Updated")
+    if user.is_authenticated():
+        user = User.objects.get(pk=user_id)
 
-    return render(request, 'taskManager/profile.html', {'user': user})
+        if request.method == 'POST':
+            form = ProfileForm(request.POST, request.FILES)
+            if form.is_valid():
+                print("made it!")
+                if request.POST.get('username') != user.username:
+                    user.username = request.POST.get('username')
+                if request.POST.get('first_name') != user.first_name:
+                    user.first_name = request.POST.get('first_name')
+                if request.POST.get('last_name') != user.last_name:
+                    user.last_name = request.POST.get('last_name')
+                if request.POST.get('email') != user.email:
+                    user.email = request.POST.get('email')
+                if request.POST.get('password'):
+                    user.set_password(request.POST.get('password'))
+                if request.FILES:
+                    user.userprofile.image = store_uploaded_file(user.username
+                    + "." + request.FILES['picture'].name.split(".")[-1], request.FILES['picture'])
+                    user.userprofile.save()
+                user.save()
+                messages.info(request, "User Updated")
+
+        return render(request, 'taskManager/profile.html', {'user': user})
 
 # A8: Cross Site Request Forgery (CSRF)
